+++ Changelog.md
@@ -7,6 +7,7 @@ Enhancements:
   to the named file. (Myron Marston)
 * Backport support for `skip` in metadata to skip execution of an example.
   (Xavier Shay, #1472)
+* Add `Pathname` support for setting all output streams. (Aaron Kromer)
 
 Deprecations:
 
@@ -57,6 +58,7 @@ Deprecations:
   `#color`, `#color=` and `#color_enabled? output`. (Jon Rowe)
 * Deprecate `#filename_pattern` in favour of `#pattern`. (Jon Rowe)
 * Deprecate `#backtrace_cleaner` in favour of `#backtrace_formatter`. (Jon Rowe)
+* Deprecate mutating `RSpec::Configuration#formatters`. (Jon Rowe)
 
 Bug Fixes: